[toolbox-sca-13] Check and alter feature values by name - Also performed code cleanup, changed api slightly to prevent user error (SequenceFactory requires a FormatterMode).

+++ toolbox-sca/src/main/java/org/haedus/soundchange/Main.java
@@ -95,7 +95,7 @@ public class Main {
 					"If running in 'standard' mode, only provide the path of a rule file.");
 		}
 
-		double elapsedTime = (double) System.nanoTime() - startTime;
+		double elapsedTime = System.nanoTime() - startTime;
 		double time = elapsedTime / StrictMath.pow(10.0,9.0);
 		LOGGER.info("Finished in {} seconds", time);
 	}
